Moved TradingView's Advanced Real-Time Chart into its own folder.
