+++ src/renderer/src/components/tradingview/AdvancedRealTimeWidget/AdvancedRealTimeWidget.css
