+++ src/renderer/src/components/tradingview/AdvancedRealTimeWidget/AdvancedRealTimeWidget.tsx
@@ -2,10 +2,9 @@ import "./AdvancedRealTimeWidget.css";
 
 import useTabKeys from '@renderer/hook/useTabKeys';
 import useTheme from '@renderer/hook/useTheme';
-import { useEffect, useRef, memo, MutableRefObject, ReactNode, useContext } from 'react';
+import { useEffect, useRef, memo, MutableRefObject, ReactNode, useState } from 'react';
 import { ASSETS } from "@renderer/assets/assets";
-import StyledIcon from "../StyledIcon/StyledIcon";
-import { ModalContext } from "@renderer/context/ModalContext";
+import StyledIcon from "@renderer/components/StyledIcon/StyledIcon";
 
 
 type Props = {
@@ -24,7 +23,9 @@ function AdvancedRealTimeWidget(props: Props): ReactNode {
 
   const {activeTheme} = useTheme();
   const {formatKey} = useTabKeys();
-  const {openModal} = useContext(ModalContext);
+  const {theme} = useTheme();
+
+  const [fullscreen, setFullscreen] = useState<boolean>(false);
 
     // THIS IS THE EMBED CODE AS GENERATED BY TRADINGVIEW, MAY NOT BE OPTIMAL!
   useEffect(() => {
@@ -51,11 +52,19 @@ function AdvancedRealTimeWidget(props: Props): ReactNode {
   
   return (
     <div
-      className="tradingview-widget-container w-100 h-100" 
       ref={container}
       key={formatKey(
         "trading-view-real-time-chart-" + pExchange + "_" + pTicker + "_" + activeTheme
       )}
+      style={fullscreen ? {
+        position:"fixed",
+        top: "0px",
+        left: "0px",
+        width: "100%",
+        height: "100%",
+        zIndex: 2
+      } : {}}
+      {...theme("ambient-bgc", "tradingview-widget-container w-100 h-100")}
     >
       <div 
         className="tradingview-widget-container__widget" 
@@ -74,12 +83,7 @@ function AdvancedRealTimeWidget(props: Props): ReactNode {
         {pAllowFullscreen && (<StyledIcon
           iconSize="advanced-real-time-widget-fullscreen"
           src={ASSETS.icons.action.fullscreen.black}
-          onClick={() => openModal(
-            <AdvancedRealTimeWidget
-              {...props}
-              allowFullscreen={false}
-            />
-          )}
+          onClick={() => pAllowFullscreen && setFullscreen(!fullscreen)}
         />)}
       </div>
     </div>
